[v0.11.1] Fix controller setup - add schemas based on CRDs available
In #233 an enhancement was done to KServe controllers to watch resources based on available CRDs.

A similar change in the setup of the manager was overlooked: it is also needed to add schemas to the manager based on available CRDs rather than only the values in the inferenceservice-config ConfigMap. This would keep both manager setup and controller setup in sync with regards schemas and watches around the CRDs.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>